requirements.txt
torch
transformers
inseq
tqdm
sacrebleu
pandas
unbabel-comet
datasets
numpy
